test: add default cart repository getItems tests

+++ app/src/testFixtures/java/com/littlelemon/application/database/cart/FakeCartDao.kt
@@ -4,7 +4,6 @@ import com.littlelemon.application.database.cart.models.CartItemDetails
 import com.littlelemon.application.database.cart.models.CartItemEntity
 import com.littlelemon.application.database.menu.models.DishWithCategories
 import com.littlelemon.application.menu.utils.DishGenerator
-import io.github.serpro69.kfaker.faker
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlin.random.Random
@@ -14,13 +13,8 @@ import kotlin.uuid.ExperimentalUuidApi
 //       and pass in a emptyList to seed with no entries.
 @OptIn(ExperimentalUuidApi::class)
 class FakeCartDao(
-    initialItems: List<CartItemEntity>? = null,
-    private val throwError: Boolean = false
-) :
-    CartDao {
-
-    private val faker = faker {}
-
+    initialItems: List<CartItemEntity>? = null, private val throwError: Boolean = false
+) : CartDao {
     private var database = mutableListOf<CartItemEntity>()
     val dishes = mutableListOf<DishWithCategories>()
 
@@ -43,34 +37,32 @@ class FakeCartDao(
 
 
     override suspend fun upsertCartItem(cartItem: CartItemEntity) {
-        if (throwError)
-            throw IllegalArgumentException()
-        if (cartItem in database)
-            database.remove(cartItem)
+        if (throwError) throw IllegalArgumentException()
+        if (cartItem !in database) {
+            val dish = DishGenerator.generateDishWithCategories(1, 1).first().first
+            dishes.add(dish.copy(dish = dish.dish.copy(dishId = cartItem.dishId)))
+        }
+        if (cartItem in database) database.remove(cartItem)
         database.add(cartItem)
     }
 
     override suspend fun removeCartItem(dishId: String) {
-        if (throwError)
-            throw IllegalArgumentException()
+        if (throwError) throw IllegalArgumentException()
         database = database.filter { it.dishId != dishId }.toMutableList()
     }
 
     override fun getAllCartItems(): Flow<List<CartItemDetails>> = flow {
-        if (throwError)
-            throw IllegalArgumentException()
+        if (throwError) throw IllegalArgumentException()
         emit(database.map {
             CartItemDetails(
                 it,
                 dishes.find { (dish) -> dish.dishId == it.dishId }
-                    ?: DishGenerator.generateDishWithCategories(1, 1).first().first
-            )
+                    ?: DishGenerator.generateDishWithCategories(1, 1).first().first)
         })
     }
 
     override fun clearCartItems() {
-        if (throwError)
-            throw IllegalArgumentException()
+        if (throwError) throw IllegalArgumentException()
         database.clear()
     }
